services/metadata-ingestion/pyproject.toml
[project]
name = "metadata-ingestion"
version = "0.1.0"
description = "Automated metadata ingestion service for DataHub"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "OpenDataGov Team"}
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "kafka-python>=2.0.2",
    "httpx>=0.25.0",
    "odg-core>=0.1.0",
    "pydantic>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.282",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/metadata_ingestion"]
